V 1.0 🥳
Everything except textures and planet cirlces which they orbit around is done.
There is still so much to be done on this project and it has much more potential in the future - better lighting, textures, planet models, gui, antialiasing, and so on...
We'll see about updates on this project later. For now it looks nice and clean (can't really say the same thing about code tho).